+++ lib/bamboohelper/misc.rb
@@ -1,8 +1,0 @@
-require 'rest-client'
-module LitaBambooHelper
-  module Misc
-    def get_request
-
-    end
-  end #module misc
-end#module