+++ lib/lita-bamboo.rb
@@ -4,7 +4,6 @@ Lita.load_locales Dir[File.expand_path(
   File.join("..", "..", "locales", "*.yml"), __FILE__
 )]
 
-require "bamboohelper/misc"
 require "bamboohelper/plans"
 require "lita/handlers/bamboo"
 require "lita"