+++ spec/lita/handlers/bamboo_spec.rb
@@ -4,11 +4,16 @@ describe Lita::Handlers::Bamboo, lita_handler: true do
   before do
     registry.config.handlers.bamboo.url = 'https://bamboo.entertainment.com/rest/api/latest'
     registry.config.handlers.bamboo.verify_ssl = false
+    registry.config.handlers.bamboo.user = 'dummy'
+    registry.config.handlers.bamboo.password = 'pass'
   end
   it do
     is_expected.to route_command('bamboo list project').to(:cmd_list_projects)
     is_expected.to route_command('bamboo list project AG plans').to(:cmd_list_project_plans)
     is_expected.to route_command('bamboo list plan AG-FPMWIL results limit 1').to(:cmd_list_plan_results)
+    is_expected.to route_command('bamboo list queue').to(:cmd_list_queue)
+    is_expected.to route_command('bamboo queue PM-RVMGEMSET').to(:cmd_queue_plan)
+    is_expected.to route_command('bamboo dequeue PM-RVMGEMSET-8').to(:cmd_dequeue_plan)
   end
 
   describe '#get all project list' do
@@ -44,6 +49,27 @@ describe Lita::Handlers::Bamboo, lita_handler: true do
       send_command('bamboo list plan AG-FPMWIL results limit 2')
       puts replies.last
       #expect(replies.last).to match(/repositoryPath/)
+    end
+  end
+
+  describe '#get build queue' do
+    it 'list build queue' do
+      send_command('bamboo list queue')
+      puts replies.last
+    end
+  end
+
+  describe '#queue plan for build' do
+    it 'queue plan' do
+      send_command('bamboo queue PM-RVMGEMSET')
+      puts replies.last
+    end
+  end
+
+  describe '#dequeue plan from build' do
+    it 'dequeue plan' do
+      send_command('bamboo dequeue PORJ-PLAN-8')
+      puts replies.last
     end
   end
 end